Test script with layer-by-layer forward and smooth
Add normalization in forward_and_smooth, the result should match the
previous Pytorch wrapper model.
Only need to add a few more smoothing method and should be ready for testing.
ToTensor() in torchvision.transforms will automatically transpose
channels, need to take care of it if the input is already channel-first.